Develop

+++ client/ayon_unreal/plugins/load/load_geometrycache_abc.py
@@ -6,11 +6,14 @@ from ayon_core.pipeline import (
     get_representation_path,
     AYON_CONTAINER_ID
 )
+from ayon_core.lib import EnumDef
 from ayon_unreal.api import plugin
 from ayon_unreal.api.pipeline import (
-    AYON_ASSET_DIR,
     create_container,
     imprint,
+    has_asset_directory_pattern_matched,
+    format_asset_directory,
+    UNREAL_VERSION
 )
 
 import unreal  # noqa
@@ -25,18 +28,62 @@ class PointCacheAlembicLoader(plugin.Loader):
     icon = "cube"
     color = "orange"
 
-    root = AYON_ASSET_DIR
+    abc_conversion_preset = "maya"
+    loaded_asset_dir = "{folder[path]}/{product[name]}"
+
+    @classmethod
+    def apply_settings(cls, project_settings):
+        super(PointCacheAlembicLoader, cls).apply_settings(project_settings)
+        # Apply import settings
+        unreal_settings = project_settings.get("unreal", {})
+        if unreal_settings.get("abc_conversion_preset", cls.abc_conversion_preset):
+            cls.abc_conversion_preset = unreal_settings.get(
+                "abc_conversion_preset", cls.abc_conversion_preset)
+        if unreal_settings.get("loaded_asset_dir", cls.loaded_asset_dir):
+            cls.loaded_asset_dir = unreal_settings.get(
+                    "loaded_asset_dir", cls.loaded_asset_dir)
+
+    @classmethod
+    def get_options(cls, contexts):
+        return [
+            EnumDef(
+                "abc_conversion_preset",
+                label="Alembic Conversion Preset",
+                items={
+                    "custom": "custom",
+                    "maya": "maya"
+                },
+                default=cls.abc_conversion_preset
+            )
+        ]
 
     @staticmethod
     def get_task(
         filename, asset_dir, asset_name, replace,
-        frame_start=None, frame_end=None
+        frame_start=None, frame_end=None, loaded_options=None
     ):
         task = unreal.AssetImportTask()
         options = unreal.AbcImportSettings()
         gc_settings = unreal.AbcGeometryCacheSettings()
         conversion_settings = unreal.AbcConversionSettings()
         sampling_settings = unreal.AbcSamplingSettings()
+        abc_conversion_preset = loaded_options.get("abc_conversion_preset")
+        if abc_conversion_preset == "maya":
+            if UNREAL_VERSION.major >= 5 and UNREAL_VERSION.minor >= 4:
+                conversion_settings = unreal.AbcConversionSettings(
+                    preset=unreal.AbcConversionPreset.MAYA)
+            else:
+                conversion_settings = unreal.AbcConversionSettings(
+                    preset=unreal.AbcConversionPreset.CUSTOM,
+                    flip_u=False, flip_v=True,
+                    rotation=[90.0, 0.0, 0.0],
+                    scale=[1.0, -1.0, 1.0])
+        else:
+            conversion_settings = unreal.AbcConversionSettings(
+                preset=unreal.AbcConversionPreset.CUSTOM,
+                flip_u=False, flip_v=True,
+                rotation=[-90.0, 0.0, 180.0],
+                scale=[100.0, 100.0, 100.0])
 
         task.set_editor_property('filename', filename)
         task.set_editor_property('destination_path', asset_dir)
@@ -47,15 +94,10 @@ class PointCacheAlembicLoader(plugin.Loader):
 
         options.set_editor_property(
             'import_type', unreal.AlembicImportType.GEOMETRY_CACHE)
+        options.sampling_settings.frame_start = frame_start
+        options.sampling_settings.frame_end = frame_end
 
         gc_settings.set_editor_property('flatten_tracks', False)
-
-        conversion_settings.set_editor_property('flip_u', False)
-        conversion_settings.set_editor_property('flip_v', True)
-        conversion_settings.set_editor_property(
-            'scale', unreal.Vector(x=100.0, y=100.0, z=100.0))
-        conversion_settings.set_editor_property(
-            'rotation', unreal.Vector(x=-90.0, y=0.0, z=180.0))
 
         if frame_start is not None:
             sampling_settings.set_editor_property('frame_start', frame_start)
@@ -71,17 +113,27 @@ class PointCacheAlembicLoader(plugin.Loader):
 
     def import_and_containerize(
         self, filepath, asset_dir, asset_name, container_name,
-        frame_start, frame_end
+        frame_start, frame_end, loaded_options=None, asset_path=None
     ):
-        unreal.EditorAssetLibrary.make_directory(asset_dir)
-
-        task = self.get_task(
-            filepath, asset_dir, asset_name, False, frame_start, frame_end)
+        task = None
+        if asset_path:
+            loaded_asset_dir = unreal.Paths.split(asset_path)[0]
+            task = self.get_task(
+                filepath, loaded_asset_dir, asset_name, True, frame_start, frame_end, loaded_options)
+        else:
+            if not unreal.EditorAssetLibrary.does_asset_exist(
+                f"{asset_dir}/{asset_name}"):
+                    task = self.get_task(
+                        filepath, asset_dir, asset_name, False,
+                        frame_start, frame_end, loaded_options
+                    )
 
         unreal.AssetToolsHelpers.get_asset_tools().import_asset_tasks([task])
 
-        # Create Asset Container
-        create_container(container=container_name, path=asset_dir)
+        if not unreal.EditorAssetLibrary.does_asset_exist(
+            f"{asset_dir}/{container_name}"):
+                # Create Asset Container
+                create_container(container=container_name, path=asset_dir)
 
     def imprint(
         self,
@@ -92,7 +144,7 @@ class PointCacheAlembicLoader(plugin.Loader):
         representation,
         frame_start,
         frame_end,
-        product_type,
+        product_type
     ):
         data = {
             "schema": "ayon:container-2.0",
@@ -109,7 +161,7 @@ class PointCacheAlembicLoader(plugin.Loader):
             "folder_path": folder_path,
             # TODO these should be probably removed
             "family": product_type,
-            "asset": folder_path,
+            "asset": folder_path
         }
         imprint(f"{asset_dir}/{container_name}", data)
 
@@ -131,21 +183,19 @@ class PointCacheAlembicLoader(plugin.Loader):
         # Create directory for asset and Ayon container
         folder_entity = context["folder"]
         folder_path = folder_entity["path"]
-        folder_name = folder_entity["name"]
         folder_attributes = folder_entity["attrib"]
 
         suffix = "_CON"
-        asset_name = f"{folder_name}_{name}" if folder_name else f"{name}"
-        version = context["version"]["version"]
-        # Check if version is hero version and use different name
-        if version < 0:
-            name_version = f"{name}_hero"
-        else:
-            name_version = f"{name}_v{version:03d}"
+        path = self.filepath_from_context(context)
+        ext = os.path.splitext(path)[-1].lstrip(".")
+        asset_root, asset_name = format_asset_directory(
+            name, context, self.loaded_asset_dir, extension=ext
+        )
 
         tools = unreal.AssetToolsHelpers().get_asset_tools()
         asset_dir, container_name = tools.create_unique_asset_name(
-            f"{self.root}/{folder_name}/{name_version}", suffix="")
+            asset_root, suffix=f"_{ext}")
+
 
         container_name += suffix
 
@@ -156,14 +206,25 @@ class PointCacheAlembicLoader(plugin.Loader):
         # one, otherwise Unreal will not import it
         if frame_start == frame_end:
             frame_end += 1
-
+        asset_path = has_asset_directory_pattern_matched(
+            asset_name, asset_dir, name, extension=ext)
         if not unreal.EditorAssetLibrary.does_directory_exist(asset_dir):
-            path = self.filepath_from_context(context)
+            unreal.EditorAssetLibrary.make_directory(asset_dir)
+        loaded_options = {
+            "abc_conversion_preset": options.get(
+                "abc_conversion_preset", self.abc_conversion_preset)
+        }
+        self.import_and_containerize(
+            path, asset_dir, asset_name, container_name,
+            frame_start, frame_end,
+            loaded_options, asset_path=asset_path
+        )
 
-            self.import_and_containerize(
-                path, asset_dir, asset_name, container_name,
-                frame_start, frame_end)
-
+        if asset_path:
+            unreal.EditorAssetLibrary.rename_asset(
+                f"{asset_path}",
+                f"{asset_dir}/{asset_name}.{asset_name}"
+            )
         self.imprint(
             folder_path,
             asset_dir,
@@ -174,7 +235,6 @@ class PointCacheAlembicLoader(plugin.Loader):
             frame_end,
             context["product"]["productType"]
         )
-
         asset_content = unreal.EditorAssetLibrary.list_assets(
             asset_dir, recursive=True, include_folder=True
         )
@@ -187,37 +247,31 @@ class PointCacheAlembicLoader(plugin.Loader):
     def update(self, container, context):
         # Create directory for folder and Ayon container
         folder_path = context["folder"]["path"]
-        folder_name = context["folder"]["name"]
         product_name = context["product"]["name"]
         product_type = context["product"]["productType"]
-        version = context["version"]["version"]
         repre_entity = context["representation"]
-
+        asset_dir = container["namespace"]
         suffix = "_CON"
-        asset_name = product_name
-        if folder_name:
-            asset_name = f"{folder_name}_{product_name}"
-
-        # Check if version is hero version and use different name
-        if version < 0:
-            name_version = f"{product_name}_hero"
-        else:
-            name_version = f"{product_name}_v{version:03d}"
+        path = get_representation_path(repre_entity)
+        ext = os.path.splitext(path)[-1].lstrip(".")
+        asset_root, asset_name = format_asset_directory(
+            product_name, context, self.loaded_asset_dir, extension=ext)
         tools = unreal.AssetToolsHelpers().get_asset_tools()
         asset_dir, container_name = tools.create_unique_asset_name(
-            f"{self.root}/{folder_name}/{name_version}", suffix="")
+            asset_root, suffix=f"_{ext}")
 
         container_name += suffix
 
         frame_start = int(container.get("frame_start"))
         frame_end = int(container.get("frame_end"))
-
         if not unreal.EditorAssetLibrary.does_directory_exist(asset_dir):
-            path = get_representation_path(repre_entity)
-
-            self.import_and_containerize(
-                path, asset_dir, asset_name, container_name,
-                frame_start, frame_end)
+            unreal.EditorAssetLibrary.make_directory(asset_dir)
+        loaded_options = {
+            "abc_conversion_preset": self.abc_conversion_preset
+        }
+        self.import_and_containerize(
+            path, asset_dir, asset_name, container_name,
+            frame_start, frame_end, loaded_options)
 
         self.imprint(
             folder_path,
@@ -239,13 +293,5 @@ class PointCacheAlembicLoader(plugin.Loader):
 
     def remove(self, container):
         path = container["namespace"]
-        parent_path = os.path.dirname(path)
-
-        unreal.EditorAssetLibrary.delete_directory(path)
-
-        asset_content = unreal.EditorAssetLibrary.list_assets(
-            parent_path, recursive=False
-        )
-
-        if len(asset_content) == 0:
-            unreal.EditorAssetLibrary.delete_directory(parent_path)
+        if unreal.EditorAssetLibrary.does_directory_exist(path):
+            unreal.EditorAssetLibrary.delete_directory(path)
